success fat16 booting

+++ src/drivers/screen.c
@@ -82,8 +82,8 @@ void    print_char(char c, int col, int row, char attribute_byte)
         for (i = 1; i < MAX_ROWS; i++)
         {
             memcpy(
-                (char*)(VIDEO_ADDRESS + (i - 1) * MAX_COLS * 2),
-                (char*)(VIDEO_ADDRESS + (i * MAX_COLS * 2)),
+                (char*)((uint64_t)VIDEO_ADDRESS + (i - 1) * MAX_COLS * 2),
+                (char*)((uint64_t)VIDEO_ADDRESS + (i * MAX_COLS * 2)),
                 MAX_COLS * 2
             );
         }
